split Person, Car and Address into separate classes. Update Spring to 2.1.8.RELEASE.
