+++ src/main/java/com/mongodb/starter/models/Person.java
@@ -120,138 +120,4 @@ public class Person {
         return Objects.hash(id, firstName, lastName, age, address, createdAt, insurance, cars);
     }
 
-    public static class Car {
-        private String brand;
-        private String model;
-        private Float maxSpeedKmH;
-
-        public Car() {
-        }
-
-        public String getBrand() {
-            return brand;
-        }
-
-        public Car setBrand(String brand) {
-            this.brand = brand;
-            return this;
-        }
-
-        public String getModel() {
-            return model;
-        }
-
-        public Car setModel(String model) {
-            this.model = model;
-            return this;
-        }
-
-        public Float getMaxSpeedKmH() {
-            return maxSpeedKmH;
-        }
-
-        public Car setMaxSpeedKmH(Float maxSpeedKmH) {
-            this.maxSpeedKmH = maxSpeedKmH;
-            return this;
-        }
-
-        @Override
-        public String toString() {
-            return "Car{" + "brand='" + brand + '\'' + ", model='" + model + '\'' + ", maxSpeedKmH=" + maxSpeedKmH + '}';
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o)
-                return true;
-            if (o == null || getClass() != o.getClass())
-                return false;
-            Car car = (Car) o;
-            return Objects.equals(brand, car.brand) && Objects.equals(model, car.model) && Objects.equals(maxSpeedKmH,
-                                                                                                          car.maxSpeedKmH);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(brand, model, maxSpeedKmH);
-        }
-    }
-
-    public static class Address {
-        private int number;
-        private String street;
-        private String postcode;
-        private String city;
-        private String country;
-
-        public Address() {
-        }
-
-        public int getNumber() {
-            return number;
-        }
-
-        public Address setNumber(int number) {
-            this.number = number;
-            return this;
-        }
-
-        public String getStreet() {
-            return street;
-        }
-
-        public Address setStreet(String street) {
-            this.street = street;
-            return this;
-        }
-
-        public String getPostcode() {
-            return postcode;
-        }
-
-        public Address setPostcode(String postcode) {
-            this.postcode = postcode;
-            return this;
-        }
-
-        public String getCity() {
-            return city;
-        }
-
-        public Address setCity(String city) {
-            this.city = city;
-            return this;
-        }
-
-        public String getCountry() {
-            return country;
-        }
-
-        public Address setCountry(String country) {
-            this.country = country;
-            return this;
-        }
-
-        @Override
-        public String toString() {
-            return "Address{" + "number=" + number + ", street='" + street + '\'' + ", postcode='" + postcode + '\'' + ", city='" + city + '\'' + ", country='" + country + '\'' + '}';
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o)
-                return true;
-            if (o == null || getClass() != o.getClass())
-                return false;
-            Address address = (Address) o;
-            return number == address.number && Objects.equals(street, address.street) && Objects.equals(postcode,
-                                                                                                        address.postcode) && Objects
-                    .equals(city, address.city) && Objects.equals(country, address.country);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(number, street, postcode, city, country);
-        }
-    }
 }
